Save a SQL server, eat Naouak !
About every query was rewritten, because score calculations were made with O(n²) algorithms, so it's about 1000 times faster now (because there is 1000 entries...)
The personnal score is now calculed the same way that in the top 10 list and the queries are easy to read. Scores seems to be a bit different from what they were, but hand made verification has shown nothing wrong with my queries and I have no courage/madness/sparta to check Naouak's ones.
And now it's impossible for an user that is not logged in to push the button (even though the button was not shown, the check was not made when the HTTP query was sent)