Added lightbox for K². Every link using lightbox as a class will display a lightbox.
Daytof has been modified in consequences and is now using it.

As the lightbox uses YUI3, yui seed file and klightbox file have been added to the default header.

The lightbox is in license BSD without having to advertise the author (in this case, me and YAHOO) and so it's compatible with Karibou² license.

The lightbox works automatically on every link containing the lightbox class in body. 
If the link contain the title attribute, it will be used as a subtitle for the picture.