A long time ago, in a network far, far away...
                   Karibou : A New Hope

           It is a period of increasing load.
         Browsers, striking from Internet, are
        doing more and more queries against the
        evil PHP/MySQL server. During the load,
      Browsers managed to download data, the HTML
   code, a taggeed text with enough syntax to show
     up an entire page. Pursued by the MySQL slow
  queries and PHP slow parser, Princess Kache races
local storage aboard her HTTP transport, custodian of
 the downloaded datas that can spare SQL queries and
     restore a good load average to the server...


Ahem. Ok so the idea is that lots of queries are going to return the same thing. Instead of server side caching, the approach here is to produce as much 304 (Not Changed) status codes as possible, through the use of Etags.

The code operates very soon in the load of Karibou to avoid parsing most of the code (this makes the use of config.xml impossible), which should make the server able to make fast replies.

Basically, if a user ask a URL, this is stored in the database. If later he asks the same page, he is replied that it has not changed. Changes are detected when an application recieves POST data, but it can also be triggered manually by the application. When changes are detected, all the lines concerning the application are removed from the database. You cannot do more precise without parsing more code...

Right now it is working, but still has a major known issue : online users data are going to be very rarely updated. Something must be done, and without using the config.xml (at least on most of queries). Of course, this could be the source of lots of bug, so please test a lot !

Another problem is that the "id" of the database might overflow "pretty" fast, administators must be careful to this.

Here comes the DB schema

CREATE TABLE IF NOT EXISTS `kache` (
  `id` int(10) unsigned NOT NULL auto_increment,
  `user_id` int(10) unsigned NOT NULL,
  `app` varchar(64) character set ascii collate ascii_bin NOT NULL,
  `args` varchar(64) character set ascii collate ascii_bin NOT NULL,
  `date` timestamp NOT NULL default CURRENT_TIMESTAMP,
  PRIMARY KEY  (`id`),
  UNIQUE KEY `user_id` (`user_id`,`app`,`args`),
  KEY `date` (`date`),
  KEY `app` (`app`,`args`)
) ENGINE=MEMORY  DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;