Add a new theme system. In this system, each theme needs a theme.config file to be classified as valid. This file contains two entries so far : - the parent theme - a description - an entry to specify whether the theme is hidden from the user or not.
The themes can inherit from another theme. When they do so, they can just redefine whatever they want, it is that easy.
There is a core theme that is *always* loaded.                                                                         

When an application is loaded, the ThemeManager will find the required CSS files using the application name. You don't have to load all the CSS 
anymore for every page. It is a quite big time saver. For miniapplications, a tiny piece of JavaScript added automatically in the appjsview     
does the job for you.                                                                                                                           

A new theme called karibou2 is included. It is basically a port of the current karibou theme to the new theming system.

The ThemeManager does not provide so far a way to list the themes. It will be possible soon. Also, users can't choose their themes so far, but 
this will be possible soon too.                                                                                                                
Also, there is a very short time between the display of a miniapplication and the loading of its CSS file, but it is unnoticeable and not a    
deal-breaker considering the savings this theme system does give us.                                                                           
I also removed the support for the in-browser theme chooser : that feature can't be used, it is not enough.                                    
There are some changes in config.php, be careful.                                                                                              

This has been discussed with Naouak to be sure this fullfills the needs of CSS writers.



List of changes :
- karibou2/class/kapp => register an application for the thememanager
- karibou2/class/karibou => feeds the classloader                    
- karibou2/class/thememanager => the theme manager                   
- apps/header => loads the right CSS files                           
- website/themes/{karibou2,core} => the new themes     
- apps/default => loads the CSS files using the the appjsview