This is a simple application to search users using their phone number. It requires the following SQL function in the framework database :
CREATE FUNCTION filterNumeric(number TEXT) RETURNS text DETERMINISTIC
BEGIN
  SET @i = 1;
  SET @result = "";
  lbl1: LOOP
    SET @c = SUBSTRING(number, @i, 1);
    IF @c REGEXP ('[0-9]') THEN SET @result = CONCAT(@result, @c); END IF;
    SET @i = @i + 1;
    IF @i <= LENGTH(number) THEN ITERATE lbl1; END IF;
    LEAVE lbl1;
  END LOOP lbl1;
  RETURN @result;
END

(Be careful with the delimiter when creating this function.)

Warning : the karibou user must be allowed to execute functions for the application to work.

The application removes any non-numeric character from the query and the stored phone numbers.
It performs a LIKE on the 'stripped' phone numbers.